refactor(Jupyter kernels): Get to compile in new crate
