+++ rust/kernel-jupyter/Cargo.toml
@@ -1,14 +1,35 @@
 [package]
-name = "kernel-jupyter"
 description = "A kernel that delegates to a Jupyter kernel"
-version = "0.0.0"
 edition = "2021"
+name = "kernel-jupyter"
+version = "0.0.0"
+
+[features]
+default = ["zmq/vendored"]
 
 [dependencies]
+chrono = "0.4.19"
+codec-ipynb = {path = "../codec-ipynb", version = "0.0.0"}
+defaults = "0.2.0"
 dirs = "4.0.0"
-kernel = { path = "../kernel", version = "0.0.0" }
+glob = "0.3.0"
+hex = "0.4.3"
+hmac = "0.11.0"
+kernel = {path = "../kernel", version = "0.0.0"}
+key-utils = {path = "../key-utils", version = "0.0.0"}
 once_cell = "1.8.0"
+path-slash = "0.1.4"
+portpicker = "0.1.1"
+reqwest = "0.11.6"
+serde = "1.0.130"
+serde_json = "1.0.68"
+serde_with = "1.11.0"
+sha2 = "0.9.8"
+tokio = "1.13.0"
+tracing = "0.1.29"
+uuid-utils = {path = "../uuid-utils", version = "0.0.0"}
+zmq = {version = "0.9.2", default-features = false}
 
 [dev-dependencies]
-tokio = { version = "1.12.0", features = ["rt", "macros"] }
-test-utils = { path = "../test-utils", version = "0.0.0" }
+test-utils = {path = "../test-utils", version = "0.0.0"}
+tokio = {version = "1.12.0", features = ["full"]}
