+++ rust/kernel-jupyter/src/messages.rs
@@ -1,8 +1,9 @@
-use crate::utils::uuids;
 use defaults::Defaults;
-use eyre::{bail, eyre, Result};
 use hmac::Mac;
-use serde::{de::DeserializeOwned, Deserialize, Serialize};
+use kernel::{
+    eyre::{bail, eyre, Result},
+    serde::{de::DeserializeOwned, Deserialize, Serialize},
+};
 use serde_json::json;
 use std::any::type_name;
 use zmq::Socket;
@@ -13,6 +14,7 @@ pub type HmacSha256 = hmac::Hmac<sha2::Sha256>;
 ///
 /// This list is from https://jupyter-client.readthedocs.io/en/stable/messaging.html.
 #[derive(Debug, Clone, Deserialize, Serialize)]
+#[serde()]
 #[allow(non_camel_case_types)]
 pub enum JupyterMessageType {
     // Messages on the shell (ROUTER/DEALER) channel
@@ -71,7 +73,11 @@ pub struct JupyterMessageHeader {
     pub(crate) msg_type: JupyterMessageType,
 
     /// A unique identifier for the message
-    #[def = "uuids::generate(uuids::Family::Generic)"]
+    ///
+    /// This uses `uuid_utils` to generate the id, but because a `msg_id`
+    /// generated by the Jupyter kernel may be longer than the 23 chars
+    /// us a `String` rather than `Uuid` type.
+    #[def = "uuid_utils::generate(\"me\").to_string()"]
     pub(crate) msg_id: String,
 
     /// A unique identifier for the kernel session
@@ -250,6 +256,7 @@ pub struct JupyterStatus {
 /// See https://jupyter-client.readthedocs.io/en/stable/messaging.html#general-message-format.
 /// Some of the below documentation is copied from there.
 #[derive(Debug, Clone, Deserialize, Serialize)]
+#[serde()]
 pub struct JupyterMessage {
     /// ZeroMQ socket identities
     pub(crate) identities: Vec<String>,
